requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
peft>=0.4.0
datasets>=2.10.0
evaluate>=0.4.0
scikit-learn>=1.2.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=9.5.0
numpy>=1.24.0
faiss-gpu>=1.7.2
tqdm>=4.65.0
sentencepiece>=0.1.99
protobuf>=3.20.0
